test: Moved mock aws bedrock server creation to `test/lib/aws-server-stubs/index.js`
