+++ test/lib/aws-server-stubs/ai-server/http2.js
@@ -338,7 +338,3 @@ function encodeChunks(chunks) {
     }
   }).pause()
 }
-
-module.exports.internals = {
-  encodeChunks
-}
