+++ test/versioned/aws-sdk-v3/common.js
@@ -16,8 +16,6 @@ const { match } = require('../../lib/custom-assertions')
 const assert = require('node:assert')
 const SEGMENT_DESTINATION = TRANS_SEGMENT
 const helper = require('../../lib/agent_helper')
-const fs = require('node:fs')
-const path = require('node:path')
 
 function checkAWSAttributes({ trace, segment, pattern, markedSegments = [] }) {
   const expectedAttrs = {
@@ -190,25 +188,6 @@ function afterEach(ctx) {
   })
 }
 
-/**
- * Determines whether to use the http or http2 mock server
- * given the `@aws-sdk/client-bedrock-runtime` package
- * version.
- * @param {string} rootPath The root path to the `node_modules` folder with AWS Bedrock `package.json`.
- * @returns {object} The mock AWS Bedrock server, http or http2.
- */
-function getAiResponseServer(rootPath = __dirname) {
-  const bedrockPackagePath = path.join(rootPath, '/node_modules/@aws-sdk/client-bedrock-runtime/package.json')
-  const semver = require('semver')
-  const { version: pkgVersion } = JSON.parse(
-    fs.readFileSync(bedrockPackagePath)
-  )
-  if (semver.gte(pkgVersion, '3.798.0')) {
-    return require('../../lib/aws-server-stubs/ai-server/http2')
-  }
-  return require('../../lib/aws-server-stubs/ai-server')
-}
-
 module.exports = {
   afterEach,
   assertChatCompletionSummary,
@@ -221,6 +200,5 @@ module.exports = {
   SEGMENT_DESTINATION,
   checkAWSAttributes,
   getMatchingSegments,
-  checkExternals,
-  getAiResponseServer
+  checkExternals
 }
